requirements.txt
django
django-crispy-forms
numpy
pandasql
requests
bs4